+++ crates/ruma-events/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [unreleased]
 
+Improvements:
+
+- Add unstable support for the `is_animated` flag for images, according to MSC4230.
+
 # 0.30.0
 
 Breaking changes: